chore: move reset button functionality to input field
